UI, MC Gen, and DATA
The python files run the user interface and the MC generator as
suggested by their names. The data files are pre-run results from the MC
generator and the information will be visible in the UI. Note that the
files provided in the Growth folder are not used by the UI and are there
for added information if the user wants it.