Big UI Revision
The code for the User Interface has been cleaned up. Changed the
functions to find path directory to match that of the Monte Carlo
generator. Eliminated imports of unused modules, greatly increasing
startup speed. Also made a new accompanying module that holds the
functions for running a single model and plotting the results. Lastly,
the number of fields specified is now the total number of fields, not
the number of fields which will become dominant at late times.